fixup: move my TestRegistry type to Runner in runner.rs
